+++ clientefs.h
@@ -1,0 +1,10 @@
+struct cliente *criar_cliente(char *nome);
+void set_cliente_next(struct cliente *cliente_atual,struct cliente *next);
+void set_cliente_prev(struct cliente *cliente_atual,struct cliente *prev);
+char *get_cliente_nome(struct cliente *cliente_atual);
+char *get_cliente_cpf(struct cliente *cliente_atual);
+char get_cliente_genero(struct cliente *cliente_atual);
+struct data get_cliente_data_n(struct cliente *cliente_atual);
+struct tag **get_cliente_tags(struct cliente *cliente_atual);
+void inserir_cliente(struct cidade *cidade_base,struct cliente *cliente_novo);
+void listar_clientes(struct cidade *cidade_atual);
