+++ passeio.h
@@ -1,0 +1,13 @@
+struct passeio *criar_passeio(int id);
+void set_passeio_id(struct passeio *passeio_atual,int id);
+void set_passeio_agenda(struct passeio *passeio_atual,int ano,int mes,int dia, int hora,int min);
+void set_passeio_ativo(struct passeio *paseio_atual,int estado);
+void set_passeio_cliente(struct passeio *passeio_atual, struct cliente *cliente_atual,int index);
+struct cliente **get_passeio_clientes(struct passeio *passeio_atual);
+void set_passeio_next(struct passeio *passeio_atual, struct passeio *prox);
+void set_passeio_prev(struct passeio *passeio_atual, struct passeio *prev);
+int get_passeio_id(struct passeio *passeio_atual);
+struct data get_passeio_agenda(struct passeio *passeio_atual);
+int get_passeio_estado(struct passeio *passeio_atual);
+void inserir_passeio(struct roteiro *roteiro_base,struct passeio *passeio_novo);
+void inserir_cliente_passeio(struct passeio *passeio_atual,struct cliente *cliente_atual);
